+++ biobb_structure_checking/commands/altloc.py
@@ -26,17 +26,21 @@ def check(strcheck):
         print(rid)
         strcheck.summary['altloc'][rid] = {}
         fix_data['altlocs'][res] = sorted(alt_loc_res[res][0].child_dict)
+        alts = set()
         for atm in alt_loc_res[res]:
             strcheck.summary['altloc'][rid][atm.id] = []
             alt_str = f"  {atm.id:4}"
             for alt in sorted(atm.child_dict):
+                alts.add(alt)
                 alt_str += f" {alt} ({atm.child_dict[alt].occupancy:4.2f})"
                 strcheck.summary['altloc'][rid][atm.id].append({
                     'loc_label': alt,
                     'occupancy': atm.child_dict[alt].occupancy
                 })
             print(alt_str)
-
+        if 'A' not in alts:
+            print(cts.MSGS['ALTLOC_NO_A'])
+            strcheck.summary['altloc'][rid]['no_altA'] = True
     return fix_data
 
 
@@ -60,7 +64,7 @@ def fix(strcheck, opts, fix_data=None):
         strcheck.args['non_interactive'],
         set_none='All'
     )
-    input_line.add_option_all()
+    input_line.add_option_none()
     input_line.add_option_list('occup', ['occupancy'])
     input_line.add_option_list('altids', altlocs, case='upper')
     input_line.add_option_list(
@@ -78,34 +82,35 @@ def fix(strcheck, opts, fix_data=None):
     if input_option == 'error':
         return cts.MSGS['UNKNOWN_SELECTION'], select_altloc
 
-    if input_option != 'all':
-        print(f"Selecting location {select_altloc}")
-        if input_option in ('occup', 'altids'):
-            select_altloc = select_altloc.upper()
-            to_fix = {
-                res: {
-                    'ats': value,
-                    'select' : select_altloc
-                } for res, value in fix_data['alt_loc_res'].items()
-            }
-
-        elif input_option == 'resnum':
-            to_fix = {}
-            selected_rnums = {}
-            for rsel in select_altloc.split(','):
-                rnum, alt = rsel.split(':')
-                selected_rnums[rnum] = alt
-            to_fix = {
-                res: {
-                    'ats': value,
-                    'select': selected_rnums[mu.residue_num(res)]
-                }
-                for res, value in fix_data['alt_loc_res'].items()
-                if mu.residue_num(res) in selected_rnums
-            }
-        for res in to_fix:
-            strcheck.strucm.select_altloc_residue(res, to_fix[res])
-
     strcheck.summary['altloc']['selected'] = select_altloc
+    if input_option == 'none':
+        print(cts.MSGS['ALTLOC_NO_SELECTION'])
+        return False
 
+    print(f"Selecting location {select_altloc}")
+    to_fix = {}
+    if input_option in ('occup', 'altids'):
+        select_altloc = select_altloc.upper()
+        to_fix = {
+            res: {
+                'ats': value,
+                'select' : select_altloc
+            } for res, value in fix_data['alt_loc_res'].items()
+        }
+    elif input_option == 'resnum':
+        to_fix = {}
+        selected_rnums = {}
+        for rsel in select_altloc.split(','):
+            rnum, alt = rsel.split(':')
+            selected_rnums[rnum] = alt
+        to_fix = {
+            res: {
+                'ats': value,
+                'select': selected_rnums[mu.residue_num(res)]
+            }
+            for res, value in fix_data['alt_loc_res'].items()
+            if mu.residue_num(res) in selected_rnums
+        }
+    for res in to_fix:
+        strcheck.strucm.select_altloc_residue(res, to_fix[res])
     return False