+++ biobb_structure_checking/modelling/structuredata.py
@@ -1,4 +1,5 @@
 ''' Class to manage internal Structure data'''
+import sys
 import biobb_structure_checking.modelling.utils as mu
 
 
@@ -148,8 +149,6 @@ class StructureData():
             for org, fin in map_fields.items():
                 if org in self.headers:
                     self.meta[fin] = ', '.join(self.headers[org])
-                else:
-                    self.meta[fin] = 'N.A.'
 
         else:
             map_fields = {
@@ -161,17 +160,17 @@ class StructureData():
             for org, fin in map_fields.items():
                 if org in self.headers:
                     self.meta[fin] = self.headers[org]
-                else:
-                    self.meta[fin] = 'N.A.'
-            if 'resolution' not in self.headers or\
-                    not self.headers['resolution']:
-                self.meta['resolution'] = 'N.A.'
-            else:
-                self.meta['resolution'] = self.headers['resolution']
         if self.biounit:
             self.meta['biounit'] = self.biounit
-        if self.meta['entry_id'] == 'XXXX':  # Recovering PDB id for Assemblies
+        if self.meta['entry_id'] == 'XXXX' or self.meta['entry_id'] == 'User':  # Recovering PDB id for Assemblies
             self.meta['entry_id'] = self.st.id
+        # Missing fields from AF entries
+        if 'title' not in self.meta:
+            self.meta['title'] = 'N.A.'
+        if 'method' not in self.meta:
+            self.meta['method'] = 'N.A.'
+        
+        
 
     def print_headers(self) -> None:
         """
@@ -184,7 +183,7 @@ class StructureData():
             else:
                 asstxt = ""
             print(
-                f" PDB id: {self.meta['entry_id']} {asstxt}\n"
+                f" Entry id: {self.meta['entry_id']} {asstxt}\n"
                 f" Title: {self.meta['title']}\n"
                 f" Experimental method: {self.meta['method']}"
             )
@@ -216,26 +215,31 @@ class StructureData():
             else:
                 self.hetatm[mu.ORGANIC].append(res)
 
+    def _print_het_group_stats(self, res_list):
+        for res in res_list:
+            if res.get_parent().get_parent().id > 0:
+                continue
+            res_label = ''
+            if not self.no_network:
+                res_label = mu.fetch_residue_name_by_id(res.get_resname())
+                if len(self.st) > 1:
+                    res_id = f"{mu.residue_id(res, False)}/*"
+                else:
+                    res_id = mu.residue_id(res)
+                print(f"{res_id} ({res_label})")
+
     def print_hetatm_stats(self):
         '''Print statistics on HETATM'''
+
         if self.hetatm[mu.MODRES]:
             print('Modified residues found')
-            for res in self.hetatm[mu.MODRES]:
-                print(mu.residue_id(res))
+            self._print_het_group_stats(self.hetatm[mu.MODRES])
         if self.hetatm[mu.METAL]:
             print('Metal/Ion residues found')
-            for res in self.hetatm[mu.METAL]:
-                if self.no_network:
-                    print(mu.residue_id(res))
-                else:
-                    print(f"{mu.residue_id(res)} ({mu.fetch_residue_name_by_id(res.get_resname())})")
+            self._print_het_group_stats(self.hetatm[mu.METAL])
         if self.hetatm[mu.ORGANIC]:
             print('Small mol ligands found')
-            for res in self.hetatm[mu.ORGANIC]:
-                if self.no_network:
-                    print(mu.residue_id(res))
-                else:
-                    print(f"{mu.residue_id(res)} ({mu.fetch_residue_name_by_id(res.get_resname())})")
+            self._print_het_group_stats(self.hetatm[mu.ORGANIC])
 
     def _check_ca_only(self):
         ca_only = True