+++ biobb_structure_checking/modelling/utils.py
@@ -53,7 +53,7 @@ BUNIT = 2
 MODELS_MAXRMS = 15.0    # Threshold value to detect NMR models (angs)
 MODEL_TYPE_LABELS = {
     ENSM: 'Ensembl/NMR',
-    BUNIT: 'BioUnit',
+    BUNIT: 'Assembly/Biological unit',
     UNKNOWN: 'Unknown'
 }
 
@@ -294,11 +294,13 @@ def same_chain(res1, res2):
     return (res1.get_parent() == res2.get_parent()) and same_model(res1, res2)
 
 
-def seq_consecutive(res1, res2):
+def seq_consecutive(res1, res2, next_res_list):
     """
     Checks whether residues belong to the same chain and
-    are consecutive in sequences, taken from residue number
+    are consecutive in sequences, taken from residue number or next_res_list if provided
     """
+    if res1 in next_res_list:
+        return next_res_list[res1] == res2
     rnum1 = res1.id[1]
     rnum2 = res2.id[1]
     return same_chain(res1, res2) and abs(rnum1 - rnum2) == 1
@@ -664,6 +666,7 @@ def check_r_list_clashes(
     rr_list,
     clash_dist,
     atom_lists,
+    next_res_list,
     join_models=True,
     severe=True,
     get_all_contacts=False,
@@ -686,6 +689,7 @@ def check_r_list_clashes(
                 res2,
                 clash_dist,
                 atom_lists,
+                next_res_list,
                 join_models,
                 severe,
                 get_all_contacts=get_all_contacts
@@ -702,9 +706,10 @@ def check_rr_clashes(
     res2,
     clash_dist,
     atom_lists,
+    next_res_list,
     join_models=True,
     severe=True,
-    get_all_contacts=False
+    get_all_contacts=False,
 ):
     """ Check all clashes between two residues """
     clash_list = {}
@@ -733,7 +738,7 @@ def check_rr_clashes(
         for cls in atom_lists:
             if is_at_in_list(atm, atom_lists[cls], res2.get_resname()):
                 ats_list2[cls].add(atm.id)
-    if res1 != res2 and not seq_consecutive(res1, res2)\
+    if res1 != res2 and not seq_consecutive(res1, res2, next_res_list)\
             and (join_models or same_model(res1, res2)):
         for at_pair in get_all_rr_distances(res1, res2):
             at1, at2, dist2 = at_pair
@@ -788,8 +793,12 @@ def get_metal_atoms(struc, metal_ats):
         # Check for CA in modified amino acids
         if 'N' in atm.get_parent() or 'C' in atm.get_parent():
             continue
-        if atm.id in metal_ats:
-            met_list.append(atm)
+        if 'element' in dir(atm):
+            if atm.element in metal_ats:
+                met_list.append(atm)
+        else:
+            if atm.id in metal_ats:
+                met_list.append(atm)
     return met_list